admin: split transform routes
split out the transform routes in the admin server to another file for
faster compilation
